Runs Shiny app for an AQUANES site

Runs Shiny app for an AQUANES site

run_app(siteName = "haridwar", use_live_data = FALSE, mySQL_conf = NULL,
  launch.browser = TRUE, ...)

Arguments

siteName
site name for shiny app (default: "haridwar")

use_live_data
should live data be used (default: FALSE)

mySQL_conf
file path to mySQL config file (.my.cnf). Only used if parameter use_live_data is TRUE and there is no .my.cnf in the app folder for the selected site (default: NULL)

launch.browser
If true, the system's default web browser will be launched automatically after the app is started (default: TRUE)

...
further arguments passed to shiny::runApp()